requirements.txt
alembic==1.13.1
aniso8601==9.0.1
apispec==6.4.0
apispec-webframeworks==0.5.2
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-CLI==0.4.0
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
flask-restplus==0.13.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-sqlalchemy==0.30.0
mistune==3.0.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.5.1
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
referencing==0.32.1
rpds-py==0.17.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.25
Telethon==1.33.1
typing_extensions==4.9.0
Werkzeug==3.0.1
